Subtle but significant change to get 'Heat Map' working
In order to iterate several paths (all different), I needed to refresh
marker and create fresh values through move() each time. If
impplementing this in a function, there's a problem with local/global
variables, and so need to change the structure of previous functions to
carry the 'marker' argument througout. This means that when calling
move(), values are taken from the initial marker variable from the
beginning.

Also returned the contour plots to a non-function state, to make calling
its variables more simple.